This commit has to do mainly with serialization of JSON<->OWL. A first, working version of the JSON->OWL mapper is complete. The approach is described in the wiki. A few auxiliary classes have been added to deal with OWL namespaces.
1) The USE_JENA static variable was removed to allow for configuration
of this parameter that doesn't need code changes. The default behavior
is still to use JENA. The configuration is in the startup JSON config.
That's a cheap way to do it, but I've created a new issue to design this
properly:https://github.com/sharegov/opencirm/issues/7.

2) OWLIndividuals changes to add a PUT implementation for individuals.
This REST endpoint can be used to save complete description of complex
individuals such as ServiceCases.

3) A few new classes have been added to deal with mapping from short
names to full IRIs. The role is very similar to the OWLAPI's
BiDirectionalShortFormProvider interface, but more geared towards our
purpose and naming strategy.

4) Since I'm working towards an improved, better defined etc. JSON<->OWL
mapping, I've modified the OWLObjecToJson class to optionally follow to
new format. Basically, it can be configured to serialize an object
always as {"iri":"theiri"} rather than as JSON string. This is to
distinguish JSON properties that are (references to) objects from
properties that are string literals. Changes to this class are the only
ones that have some potential for regression from the current behavior,
though everything should be the same.